+++ kamehouse-ui/src/main/java/com/nicobrest/kamehouse/ui/service/SessionStatusService.java
@@ -4,12 +4,15 @@ import com.nicobrest.kamehouse.commons.model.KameHouseUser;
 import com.nicobrest.kamehouse.commons.service.KameHouseUserAuthenticationService;
 import com.nicobrest.kamehouse.commons.utils.PropertiesUtils;
 import com.nicobrest.kamehouse.ui.model.SessionStatus;
+import java.util.ArrayList;
+import java.util.List;
 import javax.servlet.http.HttpSession;
 import org.apache.commons.text.StringEscapeUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
@@ -43,6 +46,11 @@ public class SessionStatusService {
     SessionStatus sessionStatus = new SessionStatus();
     sessionStatus.setUsername(StringEscapeUtils.escapeHtml4(username));
     sessionStatus.setServer(PropertiesUtils.getHostname());
+    List<String> roles = new ArrayList<>();
+    for (GrantedAuthority authority : authentication.getAuthorities()) {
+      roles.add(authority.getAuthority());
+    }
+    sessionStatus.setRoles(roles);
     try {
       KameHouseUser kameHouseUser = kameHouseUserAuthenticationService.loadUserByUsername(username);
       sessionStatus.setFirstName(kameHouseUser.getFirstName());