+++ kamehouse-ui/src/main/webapp/js/login.js
@@ -1,8 +1,13 @@
 /** Set alert messages */
 $(document).ready(() => {
   const urlParams = new URLSearchParams(window.location.search);
+  const unauthorizedPageAccess = urlParams.get('unauthorizedPageAccess');
   const error = urlParams.get('error');
   const logout = urlParams.get('logout');
+  if (!kameHouse.core.isEmpty(unauthorizedPageAccess)) {
+    const element = document.getElementById("login-alert-group-unauthorized");
+    kameHouse.util.dom.classListRemove(element, "hidden-kh");
+  }
   if (!kameHouse.core.isEmpty(error)) {
     const element = document.getElementById("login-alert-group-error");
     kameHouse.util.dom.classListRemove(element, "hidden-kh");