+++ kamehouse-ui/src/test/java/com/nicobrest/kamehouse/ui/testutils/SessionStatusTestUtils.java
@@ -6,6 +6,7 @@ import com.nicobrest.kamehouse.commons.testutils.AbstractTestUtils;
 import com.nicobrest.kamehouse.commons.testutils.TestUtils;
 import com.nicobrest.kamehouse.commons.utils.PropertiesUtils;
 import com.nicobrest.kamehouse.ui.model.SessionStatus;
+import java.util.List;
 
 /**
  * Test data and common test methods to test SessionStatus in all layers of the application.
@@ -27,11 +28,13 @@ public class SessionStatusTestUtils extends AbstractTestUtils<SessionStatus, Obj
     assertEquals(expected.getFirstName(), returned.getFirstName());
     assertEquals(expected.getLastName(), returned.getLastName());
     assertEquals(expected.getServer(), returned.getServer());
+    assertEquals(expected.getRoles(), returned.getRoles());
   }
 
   private void initSingleTestData() {
     singleTestData = new SessionStatus();
     singleTestData.setUsername("anonymousUser");
     singleTestData.setServer(PropertiesUtils.getHostname());
+    singleTestData.setRoles(List.of("ROLE_KAMISAMA"));
   }
 }